canonicialize tool

+++ viewer.py
@@ -25,7 +25,7 @@ def fix(im):
   dat = np.array(im)
   if im.mode == "P":
     # palette image
-    pp = np.array(im.getpalette()).reshape((-1, 3))
+    pp = np.array(im.getpalette()).reshape((-1, 3)).astype(np.uint8)
     sh = dat.shape
     dat = pp[dat.flatten()].reshape(list(sh)+[3])
   return dat
